Separate renderer removed. render() method split to several methods.
